Cleanup of OpenXR module scons config
