+++ modules/openxr/extensions/platform/openxr_android_extension.cpp
@@ -30,7 +30,7 @@
 
 #include "openxr_android_extension.h"
 
-#include "../openxr_api.h"
+#include "../../openxr_api.h"
 
 #include "java_godot_wrapper.h"
 #include "os_android.h"
